Refractor code and files
[Ticket: X]
