+++ backend/src/Models/AirLine.cs
@@ -1,6 +1,5 @@
 namespace FlightApi.Models;
 using System.Text.Json.Serialization;
-using System.ComponentModel.DataAnnotations.Schema;
 public class Airline
 {
     [JsonPropertyName("name")]
