+++ backend/src/Models/Arrival.cs
@@ -1,43 +1,27 @@
   namespace FlightApi.Models;
   using System.Text.Json.Serialization;
-  using System.ComponentModel.DataAnnotations.Schema;
 public class Arrival
 {
-
     [JsonPropertyName("airport")]
     public Airport Airport { get; set; }
-
-
 
     [JsonPropertyName("scheduledTimeLocal")]
     public string ScheduledTimeLocal { get; set; }
 
-
-
     [JsonPropertyName("actualTimeLocal")]
     public string ActualTimeLocal { get; set; }
-
-
 
     [JsonPropertyName("runwayTimeLocal")]
     public string RunwayTimeLocal { get; set; }
 
-
-
     [JsonPropertyName("scheduledTimeUtc")]
     public string ScheduledTimeUtc { get; set; }
-
-
 
     [JsonPropertyName("actualTimeUtc")]
     public string ActualTimeUtc { get; set; }
 
-
-
     [JsonPropertyName("runwayTimeUtc")]
     public string RunwayTimeUtc { get; set; }
-
-
 
     [JsonPropertyName("quality")]
     public List<string> Quality { get; set; }
