+++ backend/src/Models/Flight.cs
@@ -15,34 +15,23 @@ public class Flight
     [JsonPropertyName("lastUpdateUtc")]
     public string LastUpdatedUtc { get; set; }
 
-
     [JsonPropertyName("number")]
     public string Number { get; set; }
-
 
     [JsonPropertyName("callSign")]
     public string CallSign { get; set; }
 
-
-
     [JsonPropertyName("status")]
     public string Status { get; set; }
-
-
 
     [JsonPropertyName("codeshareStatus")]
     public string CodeshareStatus { get; set; }
 
-
     [JsonPropertyName("isCargo")]
     public bool IsCargo { get; set; }
 
-
-
     [JsonPropertyName("aircraft")]
     public Aircraft Aircraft { get; set; }
-
-
 
     [JsonPropertyName("airline")]
     public Airline Airline { get; set; }
