+++ backend/src/Models/Location.cs
@@ -1,6 +1,5 @@
 namespace FlightApi.Models;
 using System.Text.Json.Serialization;
-using System.ComponentModel.DataAnnotations.Schema;
  public class Location
 {
     [JsonPropertyName("lat")]
